VAGOV-2270: move and rename mapboxToken.js file
